Refactor utility scripts
